--- data/cycloguanil_ic50.csv
@@ -1,1 +1,0 @@
-﻿-6.741,-5.955,-5.562,0,-6.548,-6.282,-3.706,-1.5,-6.049,-5.778,-4.947,-3.494,-6.663,-5.925,-4.395,-2.85
