--- data/ogbunugafor_drugless.csv
@@ -1,1 +1,0 @@
-1.398,1.275,1.227,0,1.37,1.375,1.397,1.219,1.119,1.184,1.306,1,1.273,1.282,1.45,1.25
