--- data/pyrimethamine_ic50.csv
@@ -1,1 +1,0 @@
--6.286,-5.812,-4.239,0,-6.046,-5.774,-3.732,-3.55,-5.724,-5.491,-4.015,-4.6,-5.773,-5.624,-3.587,-3.3
