+++ fears.egg-info/PKG-INFO
@@ -1,0 +1,11 @@
+Metadata-Version: 1.2
+Name: fears
+Version: 0.9.0
+Summary: UNKNOWN
+Home-page: UNKNOWN
+Author: Eshan King
+Author-email: eshan.king@case.edu
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
+Requires-Python: >= 3.6.2, <=3.9.7
