+++ fears.egg-info/SOURCES.txt
@@ -1,0 +1,18 @@
+README.md
+setup.py
+fears/__init__.py
+fears.egg-info/PKG-INFO
+fears.egg-info/SOURCES.txt
+fears.egg-info/dependency_links.txt
+fears.egg-info/requires.txt
+fears.egg-info/top_level.txt
+fears/classes/__init__.py
+fears/classes/experiment_class.py
+fears/classes/population_class.py
+fears/utils/__init__.py
+fears/utils/dir_manager.py
+fears/utils/fitness.py
+fears/utils/landscape_plotter.py
+fears/utils/pharm.py
+fears/utils/plotter.py
+fears/utils/results_manager.py
