+++ fears.egg-info/requires.txt
@@ -1,0 +1,11 @@
+pathlib
+pandas
+numpy
+pickle
+lifelines
+networkx
+matplotlib
+cycler
+seaborn
+scipy
+labellines
