+++ fears/classes/__init__.py
@@ -1,0 +1,4 @@
+from .experiment_class import Experiment
+from .population_class import Population
+
+__all__ = ['Experiment', 'Population']
